etl/requirements.txt
requests==2.32.3
odfpy==1.4.1
python-dotenv